+++ screenlockpin.koplugin/plugin/ui/ctrl/lockscreenctrl.lua
@@ -104,7 +104,7 @@ local function reuseShowOverlay()
     logger.dbg("ScreenLockPin: clear & show lock")
     overlay:clearInput()
     overlay:setVisible(true)
-    UIManager:setDirty(overlay, "ui", overlay:getRefreshRegion())
+    UIManager:setDirty(overlay, "flashui", overlay:getRefreshRegion())
 end
 
 local function onResume()
@@ -161,7 +161,7 @@ local function showOrClearLockScreen(cause)
         on_unlock = function () unlockScreen("valid_pin") end,
         on_show_notes = showNotes,
     }
-    UIManager:show(overlay, "ui", overlay:getRefreshRegion())
+    UIManager:show(overlay, "flashui", overlay:getRefreshRegion())
 end
 
 return {