+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - [#8](https://github.com/oleasteo/koreader-screenlockpin/issues/8)
   Device status text with time, battery, and frontlight. 
 
+### Fixed
+
+- [#38](https://github.com/oleasteo/koreader-screenlockpin/issues/38)
+  Flash initial lock screen panel draw to avoid a transparent background by
+  accident (only black is drawn, not white). 
+
 ## [2025.12] - 2025-12-17
 
 If everything works out, this should be the last update that you have to do by