+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-[//]: # (## [Unreleased])
+## [Unreleased]
+
+### Added
+
+- [#8](https://github.com/oleasteo/koreader-screenlockpin/issues/8)
+  Device status text with time, battery, and frontlight. 
+
+### Fixed
+
+- [#38](https://github.com/oleasteo/koreader-screenlockpin/issues/38)
+  Flash initial lock screen panel draw to avoid a transparent background by
+  accident (only black is drawn, not white). 
 
 ## [2025.12] - 2025-12-17